Rename to icu4py
The name 'icu' is taken on PyPI.
