+++ README.rst
@@ -1,18 +1,18 @@
-===
-icu
-===
+======
+icu4py
+======
 
-.. image:: https://img.shields.io/readthedocs/icu?style=for-the-badge
-   :target: https://icu.readthedocs.io/en/latest/
+.. image:: https://img.shields.io/readthedocs/icu4py?style=for-the-badge
+   :target: https://icu4py.readthedocs.io/en/latest/
 
-.. image:: https://img.shields.io/github/actions/workflow/status/adamchainz/icu/main.yml.svg?branch=main&style=for-the-badge
-   :target: https://github.com/adamchainz/icu/actions?workflow=CI
+.. image:: https://img.shields.io/github/actions/workflow/status/adamchainz/icu4py/main.yml.svg?branch=main&style=for-the-badge
+   :target: https://github.com/adamchainz/icu4py/actions?workflow=CI
 
 .. image:: https://img.shields.io/badge/Coverage-100%25-success?style=for-the-badge
-   :target: https://github.com/adamchainz/icu/actions?workflow=CI
+   :target: https://github.com/adamchainz/icu4py/actions?workflow=CI
 
-.. image:: https://img.shields.io/pypi/v/icu.svg?style=for-the-badge
-   :target: https://pypi.org/project/icu/
+.. image:: https://img.shields.io/pypi/v/icu4py.svg?style=for-the-badge
+   :target: https://pypi.org/project/icu4py/
 
 .. image:: https://img.shields.io/badge/code%20style-black-000000.svg?style=for-the-badge
    :target: https://github.com/psf/black
@@ -28,4 +28,4 @@ Bindings to the ICU (International Components for Unicode) library (ICU4C).
 Documentation
 =============
 
-Please see https://icu.readthedocs.io/.
+Please see https://icu4py.readthedocs.io/.
