+++ docs/index.rst
@@ -1,5 +1,5 @@
-icu documentation
-=================
+icu4py documentation
+====================
 
 Bindings to the `ICU (International Components for Unicode) library <https://icu.unicode.org/>`__ (ICU4C).
