+++ docs/installation.rst
@@ -7,7 +7,7 @@ Requirements
 
 Python 3.10 to 3.14 supported, including free-threaded variants from Python 3.13 onwards.
 
-Only CPython is supported at this time because icu uses its C API.
+Only CPython is supported at this time because ICU uses its C API.
 
 Installation
 ------------
@@ -28,7 +28,7 @@ Installation
 
    .. code-block:: sh
 
-      python -m pip install icu
+      python -m pip install icu4py
 
    Wheels are provided for Linux and macOS.
-   The Windows build has unresolved errors, as tracked in `Issue #11 <https://github.com/adamchainz/icu/issues/11>`__.
+   The Windows build has unresolved errors, as tracked in `Issue #11 <https://github.com/adamchainz/icu4py/issues/11>`__.
