+++ src/icu4py/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from icu._ext import MessageFormat
+from icu4py._ext import MessageFormat
 
 __all__ = [
     "MessageFormat",
